requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8

# HTTP client for broker APIs
httpx==0.25.2
aiohttp==3.9.1

# Trading APIs
alpaca-py==0.23.0

# Payment processing
stripe==7.8.1

# Data validation
pydantic==2.5.0
pydantic-settings==2.1.0

# Background tasks
celery==5.3.4
redis==5.0.1

# Logging and monitoring
loguru==0.7.2
prometheus-client==0.19.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3
uuid==1.30

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0

# Web UI
jinja2==3.1.2
aiofiles==23.2.1